// front-end/src/app/core/models/employee.model.ts
export interface Employee {
  id?: string;
  email?: string;
  firstName?: string;
  lastName?: string;
  phoneNumber?: string;
  password?: string;
  cnp?: string;
  parafa?: string;
  hierarchy?:string;
  employmentDate?: string;
  address?: string;
  education?: string;
  specialty?: string;
  experience?: string;
  type: string;
}